feat: implement inline tool confirmation to replace modal dialog
- Replace ToolConfirmationDialog with InlineToolConfirmation component
- Show tool confirmation inline within conversation flow instead of modal
- Improve UX by keeping context visible during tool confirmation
- Add auto-confirm functionality with configurable count
- Update UI with PenSquare icon for new conversations
- Fix textarea height reset after sending messages
- Show user message count in conversation list instead of total
